Drop in-band signature type selection
It is possible to invoke qrexec services directly by means of
qrexec-client-vm.  With in-band signature type selection, it would be
far too easy to omit the prefix data and accidentally sign a file with
contents that begin with "a" or "b".  The resulting signature would not
be of the file itself, but rather of the file with the first byte
skipped.  Furthermore, the first byte of the file would determine what
type of signature (armored or unarmored) is generated.

Instead, require the user to explicitly choose between armored and
unarmored signatures.  qubes.GpgArmorSign always produces an armored
signature, while qubes.GpgBinarySign always produces a binary signature.
This avoids the need for in-band signaling, and also avoids the user
having to remember what the default behavior of qubes.GpgSign is.  This
does require duplicate qrexec policies for the two services, but this
can be solved by future improvements to qrexec.

Suggested-by: Marek Marczykowski-Górecki <marmarek@invisiblethingslab.com>